Allow finding this app when searching for "plotter"
When searching for "plotter" on Flathub, I only found KmPlot and some other apps, but in the back of my mind I knew there was a Gnome app. So I found this using a search engine. Would have been nice to find this plotter when searching for "plotter". I think adding "plotter" as a keyword to the appstream metainfo file would fix this?